+++ cogs/settings.py
@@ -1,10 +1,10 @@
 import asyncio
 
 from discord.ext import commands
-from db.dbase import DBase
 import discord
 
-from cogs.utils.messages import MessageManager, get_server_from_dm
+from db.dbase import DBase
+from cogs.utils.messages import MessageManager
 from cogs.utils.checks import is_admin
 
 
@@ -15,8 +15,12 @@ class Settings:
 
 
     @commands.command(pass_context=True)
-    async def prefix(self, ctx, new_prefix=None):
-        """Set a custom server prefix"""
+    async def setprefix(self, ctx, new_prefix):
+        """
+        Change the server's command prefix
+
+        Ex. '!prefix $'
+        """
         user = ctx.message.author
         channel = ctx.message.channel
         manager = MessageManager(self.bot, user, channel, [ctx.message])
@@ -28,10 +32,6 @@ class Settings:
             await manager.say("You must be an admin to do that.")
             return await manager.clear()
 
-        if not new_prefix:
-            await manager.say("You must provide a prefix.")
-            return await manager.clear()
-
         if len(new_prefix) > 5:
             await manager.say("Prefix must be less than 6 characters.")
             return await manager.clear()
@@ -40,3 +40,14 @@ class Settings:
             db.set_prefix(ctx.message.server.id, new_prefix)
             await manager.say("Command prefix has been changed to " + new_prefix)
             return await manager.clear()
+
+
+    @setprefix.error
+    async def settimezone_error(self, error, ctx):
+        if isinstance(error, commands.MissingRequiredArgument):
+            user = ctx.message.author
+            channel = ctx.message.channel
+            server = ctx.message.server
+            manager = MessageManager(self.bot, user, channel, [ctx.message])
+            await manager.say('Oops! You must provide a new prefix.')
+            await manager.clear()