+++ spirit.py
@@ -10,7 +10,8 @@ from cogs.events import Events
 from cogs.roster import Roster
 from cogs.help import Help
 from cogs.settings import Settings
-from cogs.misc import Misc
+from cogs.general import General
+from cogs.core import Core
 
 
 async def _prefix_callable(bot, message):
@@ -29,29 +30,8 @@ bot.add_cog(Events(bot))
 bot.add_cog(Roster(bot))
 bot.add_cog(Help(bot))
 bot.add_cog(Settings(bot))
-bot.add_cog(Misc(bot))
-
-
-@bot.event
-async def on_ready():
-    """Display startup information"""
-    print('Spirit v{}'.format(constants.VERSION))
-    print('Username: {}'.format(bot.user.name))
-    print('------')
-
-
-@bot.event
-async def on_server_join(server):
-    """Add server and it's members to database"""
-    with DBase() as db:
-        db.add_server(server.id)
-
-
-@bot.event
-async def on_server_remove(server):
-    """Remove server from database"""
-    with DBase() as db:
-        db.remove_server(server.id)
+bot.add_cog(General(bot))
+bot.add_cog(Core(bot))
 
 
 def setup_logging():